Move system prompt and related code into its own module
